Added backend framework and the persitence layer. Using H2
