--- enterprise/java/com/expenseplanner/enterprise/model/Expense.java
@@ -1,52 +1,0 @@
-package java.com.expenseplanner.enterprise.model;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-
-public class Expense {
-    private Long id;
-    private java.math.BigDecimal amount;
-    private java.time.LocalDate date;
-    private String category;
-    private String description;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public BigDecimal getAmount() {
-        return amount;
-    }
-
-    public void setAmount(BigDecimal amount) {
-        this.amount = amount;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public void setDate(LocalDate date) {
-        this.date = date;
-    }
-
-    public String getCategory() {
-        return category;
-    }
-
-    public void setCategory(String category) {
-        this.category = category;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-}
