+++ enterprise/src/main/java/com/expenseplanner/enterprise/service/ExpenseService.java
@@ -1,0 +1,12 @@
+package com.expenseplanner.enterprise.service;
+
+import java.util.List;
+import com.expenseplanner.enterprise.model.Expense;
+
+public interface ExpenseService {
+    List<Expense> getAllExpenses();
+    Expense getExpenseById(Long id);
+    Expense addExpense(Expense expense);
+    Expense updateExpense(Expense expense);
+    void deleteExpense(Long id);
+}
