+++ enterprise/src/main/resources/application.properties
@@ -1,1 +1,14 @@
-spring.application.name=enterprise
+# H2 in-memory database
+spring.datasource.url=jdbc:h2:mem:expenses_db
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+
+# settings for JPA
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+
+#H2
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-console
